+++ spec/flows/contract/tuple_spec.rb
@@ -9,18 +9,11 @@ RSpec.describe Flows::Contract::Tuple do
     it_behaves_like 'Flows::Contract with valid value',
                     value: [:a, 10]
 
-    error_message = "array element `\"a\"` with index 0 is invalid:\n" \
-                    "    must match `Symbol`\n" \
-                    "array element `\"b\"` with index 1 is invalid:\n" \
-                    '    must match `Integer`'
+    it_behaves_like 'Flows::Contract with invalid value',
+                    value: %w[a b]
 
     it_behaves_like 'Flows::Contract with invalid value',
-                    value: %w[a b],
-                    error_message: error_message
-
-    it_behaves_like 'Flows::Contract with invalid value',
-                    value: [],
-                    error_message: 'array length mismatch: must be 2, got 0'
+                    value: []
   end
 
   context 'with transforms' do
@@ -35,13 +28,7 @@ RSpec.describe Flows::Contract::Tuple do
                     value: [' aa ', ' bb '],
                     after_transform: ['aa', ' BB ']
 
-    error_message = "array element `1` with index 0 is invalid:\n" \
-                    "    must match `String`\n" \
-                    "array element `2` with index 1 is invalid:\n" \
-                    '    must match `String`'
-
     it_behaves_like 'Flows::Contract with invalid value',
-                    value: [1, 2],
-                    error_message: error_message
+                    value: [1, 2]
   end
 end